# gen/http/voting/server/encode_decode.go
// Code generated by goa v3.24.1, DO NOT EDIT.
//
// voting HTTP server encoders and decoders
//
// Command:
// $ goa gen
// github.com/linuxfoundation/lfx-v2-voting-service/api/voting/v1/design

package server

import (
	"context"
	"errors"
	"io"
	"net/http"
	"strings"

	voting "github.com/linuxfoundation/lfx-v2-voting-service/gen/voting"
	goahttp "goa.design/goa/v3/http"
	goa "goa.design/goa/v3/pkg"
)

// EncodeCreateVoteResponse returns an encoder for responses returned by the
// voting create_vote endpoint.
func EncodeCreateVoteResponse(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder) func(context.Context, http.ResponseWriter, any) error {
	return func(ctx context.Context, w http.ResponseWriter, v any) error {
		res, _ := v.(*voting.VoteResult)
		enc := encoder(ctx, w)
		body := NewCreateVoteResponseBody(res)
		w.WriteHeader(http.StatusCreated)
		return enc.Encode(body)
	}
}

// DecodeCreateVoteRequest returns a decoder for requests sent to the voting
// create_vote endpoint.
func DecodeCreateVoteRequest(mux goahttp.Muxer, decoder func(*http.Request) goahttp.Decoder) func(*http.Request) (*voting.CreateVotePayload, error) {
	return func(r *http.Request) (*voting.CreateVotePayload, error) {
		var (
			body CreateVoteRequestBody
			err  error
		)
		err = decoder(r).Decode(&body)
		if err != nil {
			if errors.Is(err, io.EOF) {
				return nil, goa.MissingPayloadError()
			}
			var gerr *goa.ServiceError
			if errors.As(err, &gerr) {
				return nil, gerr
			}
			return nil, goa.DecodePayloadError(err.Error())
		}
		err = ValidateCreateVoteRequestBody(&body)
		if err != nil {
			return nil, err
		}

		var (
			token *string
		)
		tokenRaw := r.Header.Get("Authorization")
		if tokenRaw != "" {
			token = &tokenRaw
		}
		payload := NewCreateVotePayload(&body, token)
		if payload.Token != nil {
			if strings.Contains(*payload.Token, " ") {
				// Remove authorization scheme prefix (e.g. "Bearer")
				cred := strings.SplitN(*payload.Token, " ", 2)[1]
				payload.Token = &cred
			}
		}

		return payload, nil
	}
}

// EncodeCreateVoteError returns an encoder for errors returned by the
// create_vote voting endpoint.
func EncodeCreateVoteError(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder, formatter func(ctx context.Context, err error) goahttp.Statuser) func(context.Context, http.ResponseWriter, error) error {
	encodeError := goahttp.ErrorEncoder(encoder, formatter)
	return func(ctx context.Context, w http.ResponseWriter, v error) error {
		var en goa.GoaErrorNamer
		if !errors.As(v, &en) {
			return encodeError(ctx, w, v)
		}
		switch en.GoaErrorName() {
		case "BadRequest":
			var res *voting.BadRequestError
			errors.As(v, &res)
			enc := encoder(ctx, w)
			var body any
			if formatter != nil {
				body = formatter(ctx, res)
			} else {
				body = NewCreateVoteBadRequestResponseBody(res)
			}
			w.Header().Set("goa-error", res.GoaErrorName())
			w.WriteHeader(http.StatusBadRequest)
			return enc.Encode(body)
		case "Conflict":
			var res *voting.ConflictError
			errors.As(v, &res)
			enc := encoder(ctx, w)
			var body any
			if formatter != nil {
				body = formatter(ctx, res)
			} else {
				body = NewCreateVoteConflictResponseBody(res)
			}
			w.Header().Set("goa-error", res.GoaErrorName())
			w.WriteHeader(http.StatusConflict)
			return enc.Encode(body)
		case "Forbidden":
			var res *voting.ForbiddenError
			errors.As(v, &res)
			enc := encoder(ctx, w)
			var body any
			if formatter != nil {
				body = formatter(ctx, res)
			} else {
				body = NewCreateVoteForbiddenResponseBody(res)
			}
			w.Header().Set("goa-error", res.GoaErrorName())
			w.WriteHeader(http.StatusForbidden)
			return enc.Encode(body)
		case "InternalServerError":
			var res *voting.InternalServerError
			errors.As(v, &res)
			enc := encoder(ctx, w)
			var body any
			if formatter != nil {
				body = formatter(ctx, res)
			} else {
				body = NewCreateVoteInternalServerErrorResponseBody(res)
			}
			w.Header().Set("goa-error", res.GoaErrorName())
			w.WriteHeader(http.StatusInternalServerError)
			return enc.Encode(body)
		case "ServiceUnavailable":
			var res *voting.ServiceUnavailableError
			errors.As(v, &res)
			enc := encoder(ctx, w)
			var body any
			if formatter != nil {
				body = formatter(ctx, res)
			} else {
				body = NewCreateVoteServiceUnavailableResponseBody(res)
			}
			w.Header().Set("goa-error", res.GoaErrorName())
			w.WriteHeader(http.StatusServiceUnavailable)
			return enc.Encode(body)
		case "Unauthorized":
			var res *voting.UnauthorizedError
			errors.As(v, &res)
			enc := encoder(ctx, w)
			var body any
			if formatter != nil {
				body = formatter(ctx, res)
			} else {
				body = NewCreateVoteUnauthorizedResponseBody(res)
			}
			w.Header().Set("goa-error", res.GoaErrorName())
			w.WriteHeader(http.StatusUnauthorized)
			return enc.Encode(body)
		default:
			return encodeError(ctx, w, v)
		}
	}
}

// unmarshalPollQuestionRequestBodyToVotingPollQuestion builds a value of type
// *voting.PollQuestion from a value of type *PollQuestionRequestBody.
func unmarshalPollQuestionRequestBodyToVotingPollQuestion(v *PollQuestionRequestBody) *voting.PollQuestion {
	res := &voting.PollQuestion{
		QuestionID: v.QuestionID,
		Prompt:     *v.Prompt,
		Type:       *v.Type,
	}
	res.Choices = make([]*voting.PollChoice, len(v.Choices))
	for i, val := range v.Choices {
		if val == nil {
			res.Choices[i] = nil
			continue
		}
		res.Choices[i] = unmarshalPollChoiceRequestBodyToVotingPollChoice(val)
	}

	return res
}

// unmarshalPollChoiceRequestBodyToVotingPollChoice builds a value of type
// *voting.PollChoice from a value of type *PollChoiceRequestBody.
func unmarshalPollChoiceRequestBodyToVotingPollChoice(v *PollChoiceRequestBody) *voting.PollChoice {
	res := &voting.PollChoice{
		ChoiceID:   v.ChoiceID,
		ChoiceText: *v.ChoiceText,
	}

	return res
}

// marshalVotingPollQuestionToPollQuestionResponseBody builds a value of type
// *PollQuestionResponseBody from a value of type *voting.PollQuestion.
func marshalVotingPollQuestionToPollQuestionResponseBody(v *voting.PollQuestion) *PollQuestionResponseBody {
	if v == nil {
		return nil
	}
	res := &PollQuestionResponseBody{
		QuestionID: v.QuestionID,
		Prompt:     v.Prompt,
		Type:       v.Type,
	}
	if v.Choices != nil {
		res.Choices = make([]*PollChoiceResponseBody, len(v.Choices))
		for i, val := range v.Choices {
			if val == nil {
				res.Choices[i] = nil
				continue
			}
			res.Choices[i] = marshalVotingPollChoiceToPollChoiceResponseBody(val)
		}
	} else {
		res.Choices = []*PollChoiceResponseBody{}
	}

	return res
}

// marshalVotingPollChoiceToPollChoiceResponseBody builds a value of type
// *PollChoiceResponseBody from a value of type *voting.PollChoice.
func marshalVotingPollChoiceToPollChoiceResponseBody(v *voting.PollChoice) *PollChoiceResponseBody {
	res := &PollChoiceResponseBody{
		ChoiceID:   v.ChoiceID,
		ChoiceText: v.ChoiceText,
	}

	return res
}